mptcp: move to next addr when subflow creation fail
Git-commit: 557963c383e8209b14de91bf2a0301a41b94d8c4
Message-Id: <557963c383e8209b14de91bf2a0301a41b94d8c4.1637073562.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.13-rc1

When an invalid address was announced, the subflow couldn't be created
for this address. Therefore mptcp_pm_nl_subflow_established couldn't be
invoked. Then the next addresses in the local address list didn't have a
chance to be announced.

This patch invokes the new function mptcp_pm_add_addr_echoed when the
address is echoed. In it, use mptcp_lookup_anno_list_by_saddr to check
whether this address is in the anno_list. If it is, PM schedules the
status MPTCP_PM_SUBFLOW_ESTABLISHED to invoke
mptcp_pm_create_subflow_or_signal_addr to deal with the next address in
the local address list.

Signed-off-by: Geliang Tang <geliangtang@gmail.com>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.13-rc1-mptcp-move-to-next-addr-when-subflow-creation-fail.patch