RDMA/hns: Support configuring doorbell mode of RQ and CQ
From cf8cd4ccb269dbd57c3792799d0e5251547d6734 Mon Sep 17 00:00:00 2001
Git-commit: cf8cd4ccb269dbd57c3792799d0e5251547d6734
Patch-mainline: v5.13-rc1
References: bsn#12

HIP08 supports both normal and record doorbell mode for RQ and CQ, SQ
record doorbell for userspace is also supported by the software for
flushing CQE process. As now the capability of HIP08 are exposed to the
user and are configurable, the support of normal doorbell should be added
back.

Note that, if switching to normal doorbell, the kernel will report "flush
cqe is unsupported" if modify qp to error status as the flush is based on
record doorbell.

Link: https://lore.kernel.org/r/1616840738-7866-2-git-send-email-liweihang@huawei.com
Signed-off-by: Yixian Liu <liuyixian@huawei.com>
Signed-off-by: Weihang Li <liweihang@huawei.com>
Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.13-rc1-RDMA-hns-Support-configuring-doorbell-mode-of-RQ-and.patch