mm/memcg: fix 5.10 backport of splitting page memcg
Git-commit: 002ea848d7fd3bdcb6281e75bdde28095c2cd549
References: bsn#19
Patch-mainline: v5.10.27

The straight backport of 5.12's e1baddf8475b ("mm/memcg: set memcg when
splitting page") works fine in 5.11, but turned out to be wrong for 5.10:
because that relies on a separate flag, which must also be set for the
memcg to be recognized and uncharged and cleared when freeing. Fix that.

Signed-off-by: Hugh Dickins <hughd@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.27-0207-mm-memcg-fix-5.10-backport-of-splitting-page-memcg.patch