crypto: hisilicon/sec - Fix a module parameter error
From da6503f52bf85a7da29fbd7126320658f906dd9a Mon Sep 17 00:00:00 2001
Git-commit: da6503f52bf85a7da29fbd7126320658f906dd9a
Patch-mainline: v5.13-rc1
References: bsn#12

ctx_q_num is a module parameter set by the user to specify the
number of qp queues required to create a ctx.

When the number of qp queues allocated by PF or VF is less than
the ctx_q_num, an error will be reported when ctx is initialized
in kernel mode, which leads to the problem that the registered
algorithms cannot be used.

Therefore, when PF or VF is initialized, if the number of qp queues
is not enough to create a ctx, the kernel mode cannot be used,
and there is no need to register the kernel mode algorithms.

Signed-off-by: Longfang Liu <liulongfang@huawei.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.13-rc1-crypto-hisilicon-sec-Fix-a-module-parameter-error.patch